save variables of trigger block before Reply Event flow
